Add evaluation errors to koan message.
This commit attempts to implement the second option proposed in the
discussion to issue #6:

This commit adds the evaluation errors that occur in the course
of testing a koan to the default koan message.

These errors are not added if the user has not completed the koan (i.e. there
are still "__" in the koan).

Below is an example of the type of output this change introduces:

Now meditate upon /home/vagrant/clojure-koans/src/koans/17_macros.clj:42
---------------------
Assertion failed!
Really, you don't understand recursion until you understand recursion
(= 36 (r-infix (10 + (2 * 3) + (4 * 5))))

-------------
Part-2 Error: java.lang.Long cannot be cast to clojure.lang.IFn